refactor: extract common task queue methods into BaseTaskQueue interface
Agent-Id: agent-6fd9a2be-89a2-4962-b72b-c5c7f0e7f559
Linked-Note-Id: d272e28a-4907-4548-a03b-d644218e12bf
